refactor(welcome): rename sign-up dir to signup
Align the directory name with the signup-dialog/signup__* naming used
everywhere else. Update the one external import path, mirror the test dir,
and fix the stale sign-up__form testid.
