+++ src/views/welcome/signup/form.vue
@@ -19,7 +19,7 @@ const { t } = useI18n()
 </script>
 
 <template>
-  <div data-testid="sign-up__form" class="w-full max-w-94.5 flex flex-col gap-8">
+  <div data-testid="signup__form" class="w-full max-w-94.5 flex flex-col gap-8">
     <div data-testid="social-auth" class="flex flex-col gap-4.5">
       <ui-button
         size="xl"
