+++ src/views/welcome/signup/index.vue
