+++ src/views/welcome/signup/plan-option.vue
